Confirmation for Baselines + Fit Methods
Added in confirmation for baselining - once the polynomial order has been chosen, it is applied and a confirmation message is given. If the user types anything other than just hitting Enter, it allows them to go back and try other orders (in case it didn't zero well). 
For confirmation of fit method, user must enter "yes". If they do not, allows explained syntax to alter choice to a specific method, in case they decided their original, planned method would not work well for this spectrum. 
To disable this option, user merely needs to add "-noconfirm" to the argument when running the main script.