User Interaction Updates
Now prevents the user from crashing the code by pressing Enter before a region has been completed.
Specified which fit the user is selecting a region for.
Added layers of confirmation to trapezoidal fit, so the user may restart their point selection if desired.
The prompt for comments now comes after the values have been determined for trap fits.
Fixed the issue with twopeak generation a second red plot when moving on to a trap fit.