First Draft/Working Script of a Multi-gaussian Model
This is my still fairly WIP script which convolves the spectrum to a gaussian kernel (for ease of viewing during data reduction, as well as an important function while creating the new fit), and then creates a multi-gaussian fit using that convolution as an aid (but not fit to the convolution). 
The process is not yet perfect, and needs certain fixes, and to be made more user friendly, but this reduction procedure should work as is. 
Should be called in the parser as reduce.py is, but only has a parameter for smoothing (has not been tested without smoothing - if no smooth parameter is input it defaults to 21), and galaxy number itself.