Multigauss V2 - added functionality to measure.py as well
Round 2 of changes and fixes to the multigauss routine. Squashed interesting bugs and added a few checks, in event of high error or running out of iterations.
For measure.py, the only change was allowing an overlay, if desired, using multigauss's gaussian filter. This can aid in visualization of noisier data. 
As a result, two new options have been added to reduce.py's parser: '-overlay', which causes the gaussian convolution to be shown when running measure.py, and '-multigauss', which tells it to additionally (or exclusively!) run the many gaussian fit.