First round of automation
Enter autozeros.py, which examines the data to attempt to locate the signal, and uses that information to find zeros for -gauss and -multigauss. It checks 1000km/s in either direction of center, for points that, when averaged with the two neighboring channels in either direction, has a mean > 3*rms. Through testing, this works for all but the absolute lowest S/N, worst baselined spectra. Then it plots lines fitted to the sides of the profile, and uses those to find zero points, which are then given to multigauss and gauss to choose the region to fit to. Through testing, this has so far worked. Automation is current the default in this code, and requires the -noauto argument to turn off.